+++ frontend/src/components/layout/Main.tsx
@@ -1,3 +1,4 @@
+import { AnimatePresence, motion } from "framer-motion";
 import type React from "react";
 import { memo, useMemo } from "react";
 import { NavBar } from "./NavBar";
@@ -29,9 +30,17 @@ export const Main: React.FC<MainProps> = memo(
         style={style}
       >
         {showNavBar && <NavBar />}
-        <div className="hero-content w-full">
-          <div className="max-w-6xl w-full">{children}</div>
-        </div>
+        <AnimatePresence>
+          <motion.div
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0 }}
+            transition={{ duration: 1 }}
+            className="hero-content w-full"
+          >
+            <div className="max-w-6xl w-full">{children}</div>
+          </motion.div>
+        </AnimatePresence>
       </main>
     );
   },